+++ src/roadveh_cmd.cpp
@@ -373,7 +373,7 @@ CommandCost CmdSellRoadVeh(TileIndex tile, uint32 flags, uint32 p1, uint32 p2)
 		DeleteVehicle(v);
 	}
 
-	return CommandCost(-(int32)v->value);
+	return CommandCost(-v->value);
 }
 
 struct RoadFindDepotData {
